refactor(rollout): retire v1 sglang engine; wire everything onto sglang_diffusion
Make `sglang_diffusion` (v2) the single SGLang diffusion rollout engine and
delete the legacy v1 `sglang` engine core.

- relocate the shared `_patches/` hijack suite into the v2 package
  (sglang/_patches -> sglang_diffusion/_patches) + rewrite all import paths;
  v2's native backend is the sole consumer
- delete v1 engine core (engine/config/request/response/_sample_decode/_text_fusion)
- migrate all 11 sglang recipes onto the v2 engine targets: merge the two
  duplicated v1/v2 pairs onto the v1 base (discarding twin drift), delete the
  drifted twins, rename qwen_image_sglang_diffusion -> qwen_image_sglang
- fix a dangling algorithm._target_ (diffusion_grpo.DiffusionGRPO ->
  flowgrpo.FlowGRPO) that all three prior v2 recipes carried
- fix validate_rollout_layout's multi-GPU-colocate engine-suffix gate for v2
- drop the v2-vs-v1 parity test + legacy-translator cross-check; relocate the
  patch test; update docstrings/pyproject references

Verified: static sweep clean, all 44 recipe _target_s resolve, 93 passed /
3 skipped (GPU). Parity burns for the full-finetune/separate/NFT transports vs
the trainside oracle remain a follow-up.
