+++ unirl/models/flux2_klein/pipeline.py
@@ -18,7 +18,7 @@ config.
 
 σ schedule contract
 -------------------
-The hosting engine (``TrainsideRolloutEngine`` / ``SGLangRolloutEngine``
+The hosting engine (``TrainsideRolloutEngine`` / ``SGLangDiffusionRolloutEngine``
 / ``VLLMOmniRolloutEngine``) pins ``req.sigmas`` via
 :func:`unirl.sde.runtime.ensure_req_sigmas` BEFORE calling
 ``generate(req)``; this pipeline reads ``req.sigmas`` and uses it
