+++ unirl/models/hunyuan_video/pipeline.py
@@ -12,7 +12,7 @@ the stages with the precision policy from the config.
 
 sigma schedule contract
 -----------------------
-The hosting engine (``TrainsideRolloutEngine`` / ``SGLangRolloutEngine``
+The hosting engine (``TrainsideRolloutEngine`` / ``SGLangDiffusionRolloutEngine``
 / ``VLLMOmniRolloutEngine``) pins ``req.sigmas`` via
 :func:`unirl.sde.runtime.ensure_req_sigmas` BEFORE calling
 ``generate(req)``; this pipeline reads ``req.sigmas`` and uses it
