+++ unirl/rollout/engine/sglang_diffusion/_patches/hijack.py
@@ -13,7 +13,7 @@ diffusion path never calls srt ``load_plugins()``, so the official
 the worker; ``wrap_mp_process_for_children`` propagates the install into every
 spawn child instead.
 
-Install once at the top of ``SGLangRolloutEngine.__init__`` -- BEFORE importing
+Install once when the native backend boots -- BEFORE importing
 ``DiffGenerator`` (which forces spawn at import) and before ``from_pretrained``
 spawns the scheduler. Idempotent; safe to call from both parent and child.
 """
@@ -122,54 +122,54 @@ class SglangDiffusionHijack:
         # target unavailable in this interpreter is logged and skipped by
         # _safe_apply, so partial availability (e.g. a CPU-only unit-test
         # process importing only the rollout math) never aborts the rest.
-        from unirl.rollout.engine.sglang._patches.patch_conditions import (
+        from unirl.rollout.engine.sglang_diffusion._patches.patch_conditions import (
             patch_conditions,
         )
-        from unirl.rollout.engine.sglang._patches.patch_dance import patch_dance
-        from unirl.rollout.engine.sglang._patches.patch_denoising import (
+        from unirl.rollout.engine.sglang_diffusion._patches.patch_dance import patch_dance
+        from unirl.rollout.engine.sglang_diffusion._patches.patch_denoising import (
             patch_denoising,
         )
-        from unirl.rollout.engine.sglang._patches.patch_gpu_worker import (
+        from unirl.rollout.engine.sglang_diffusion._patches.patch_gpu_worker import (
             patch_gpu_worker,
         )
-        from unirl.rollout.engine.sglang._patches.patch_grouped_dispatch import (
+        from unirl.rollout.engine.sglang_diffusion._patches.patch_grouped_dispatch import (
             patch_grouped_dispatch,
         )
-        from unirl.rollout.engine.sglang._patches.patch_latent_prep import (
+        from unirl.rollout.engine.sglang_diffusion._patches.patch_latent_prep import (
             patch_latent_prep,
         )
-        from unirl.rollout.engine.sglang._patches.patch_lora_slice_2d import (
+        from unirl.rollout.engine.sglang_diffusion._patches.patch_lora_slice_2d import (
             patch_lora_slice_2d,
         )
-        from unirl.rollout.engine.sglang._patches.patch_lora_tensors import (
+        from unirl.rollout.engine.sglang_diffusion._patches.patch_lora_tensors import (
             patch_lora_tensors,
         )
-        from unirl.rollout.engine.sglang._patches.patch_pipeline import (
+        from unirl.rollout.engine.sglang_diffusion._patches.patch_pipeline import (
             patch_pipeline,
         )
-        from unirl.rollout.engine.sglang._patches.patch_platform_device import (
+        from unirl.rollout.engine.sglang_diffusion._patches.patch_platform_device import (
             patch_platform_device,
         )
-        from unirl.rollout.engine.sglang._patches.patch_rollout_trajectory import (
+        from unirl.rollout.engine.sglang_diffusion._patches.patch_rollout_trajectory import (
             patch_rollout_trajectory,
         )
-        from unirl.rollout.engine.sglang._patches.patch_sampling_io import (
+        from unirl.rollout.engine.sglang_diffusion._patches.patch_sampling_io import (
             patch_sampling_io,
         )
-        from unirl.rollout.engine.sglang._patches.patch_scheduler import (
+        from unirl.rollout.engine.sglang_diffusion._patches.patch_scheduler import (
             patch_scheduler,
         )
-        from unirl.rollout.engine.sglang._patches.patch_sd3_lora_pipeline import (
+        from unirl.rollout.engine.sglang_diffusion._patches.patch_sd3_lora_pipeline import (
             patch_sd3_lora_pipeline,
         )
-        from unirl.rollout.engine.sglang._patches.patch_set_timesteps import (
+        from unirl.rollout.engine.sglang_diffusion._patches.patch_set_timesteps import (
             patch_set_timesteps,
         )
-        from unirl.rollout.engine.sglang._patches.patch_srt import patch_srt
-        from unirl.rollout.engine.sglang._patches.patch_vae_decode_safe import (
+        from unirl.rollout.engine.sglang_diffusion._patches.patch_srt import patch_srt
+        from unirl.rollout.engine.sglang_diffusion._patches.patch_vae_decode_safe import (
             patch_vae_decode_safe,
         )
-        from unirl.rollout.engine.sglang._patches.patch_weights_updater import (
+        from unirl.rollout.engine.sglang_diffusion._patches.patch_weights_updater import (
             patch_weights_updater,
         )
