+++ unirl/rollout/engine/sglang_diffusion/_patches/io_struct.py
