+++ unirl/rollout/engine/sglang_diffusion/_patches/patch_conditions.py
