+++ unirl/rollout/engine/sglang_diffusion/_patches/patch_latent_prep.py
