+++ unirl/rollout/engine/sglang_diffusion/_patches/patch_pipeline.py
