+++ unirl/rollout/engine/sglang_diffusion/_patches/patch_platform_device.py
