+++ unirl/rollout/engine/sglang_diffusion/_patches/patch_rollout_trajectory.py
