+++ unirl/rollout/engine/sglang_diffusion/_patches/patch_sampling_io.py
