+++ unirl/rollout/engine/sglang_diffusion/_patches/patch_vae_decode_safe.py
